uint -> uint256

+++ mocks/pumps/MockFailPump.sol
@@ -1,6 +1,5 @@
 /**
  * SPDX-License-Identifier: MIT
- *
  */
 
 pragma solidity ^0.8.17;
@@ -11,7 +10,7 @@ import "src/interfaces/pumps/IPump.sol";
  * @title Mock Pump with a failing update function
  */
 contract MockFailPump is IPump {
-    function update(uint[] calldata, bytes calldata) external pure {
+    function update(uint256[] calldata, bytes calldata) external pure {
         revert();
     }
 }
