+++ src/libraries/LibMath.sol
@@ -11,7 +11,7 @@ library LibMath {
      * @param b denominator
      * @dev Division, rounded up
      */
-    function roundUpDiv(uint a, uint b) internal pure returns (uint) {
+    function roundUpDiv(uint256 a, uint256 b) internal pure returns (uint256) {
         if (a == 0) return 0;
         return (a - 1) / b + 1;
     }
@@ -29,7 +29,7 @@ library LibMath {
      * thus we recursively call sqrt().
      *
      */
-    function nthRoot(uint a, uint n) internal pure returns (uint root) {
+    function nthRoot(uint256 a, uint256 n) internal pure returns (uint256 root) {
         assert(n > 1);
         if (a == 0) return 0;
         // Equivalent to "i % 2 == 0" but cheaper.
@@ -41,13 +41,13 @@ library LibMath {
         }
         // The scale factor is a crude way to turn everything into integer calcs.
         // Actually do ((10 ^ n) * a) ^ (1/n)
-        uint a0 = (10 ** n) * a;
+        uint256 a0 = (10 ** n) * a;
 
-        uint xNew = 10;
-        uint x;
+        uint256 xNew = 10;
+        uint256 x;
         while (xNew != x) {
             x = xNew;
-            uint t0 = x ** (n - 1);
+            uint256 t0 = x ** (n - 1);
             if (x * t0 > a0) {
                 xNew = x - (x - a0 / t0) / n;
             } else {
@@ -69,7 +69,7 @@ library LibMath {
      * based on https://github.com/transmissions11/solmate/blob/main/src/utils/FixedPointMathLib.sol
      */
 
-    function sqrt(uint a) internal pure returns (uint z) {
+    function sqrt(uint256 a) internal pure returns (uint256 z) {
         /// @solidity memory-safe-assembly
         assembly {
             let y := a // We start y at a, which will help us make our initial estimate.
@@ -134,20 +134,20 @@ library LibMath {
         }
     }
 
-    /// @notice Calculates floor(a×b÷denominator) with full precision. Throws if result overflows a uint or denominator == 0
+    /// @notice Calculates floor(a×b÷denominator) with full precision. Throws if result overflows a uint256 or denominator == 0
     /// @param a The multiplicand
     /// @param b The multiplier
     /// @param denominator The divisor
     /// @return result The 256-bit result
     /// @dev Credit to Remco Bloemen under MIT license https://xn--2-umb.com/21/muldiv
-    function mulDiv(uint a, uint b, uint denominator) internal pure returns (uint result) {
+    function mulDiv(uint256 a, uint256 b, uint256 denominator) internal pure returns (uint256 result) {
         // 512-bit multiply [prod1 prod0] = a * b
         // Compute the product mod 2**256 and mod 2**256 - 1
         // then use the Chinese Remainder Theorem to reconstruct
         // the 512 bit result. The result is stored in two 256
         // variables such that product = prod1 * 2**256 + prod0
-        uint prod0; // Least significant 256 bits of the product
-        uint prod1; // Most significant 256 bits of the product
+        uint256 prod0; // Least significant 256 bits of the product
+        uint256 prod1; // Most significant 256 bits of the product
         assembly {
             let mm := mulmod(a, b, not(0))
             prod0 := mul(a, b)
@@ -173,7 +173,7 @@ library LibMath {
 
         // Make division exact by subtracting the remainder from [prod1 prod0]
         // Compute remainder using mulmod
-        uint remainder;
+        uint256 remainder;
         assembly {
             remainder := mulmod(a, b, denominator)
         }
@@ -187,7 +187,7 @@ library LibMath {
         // Compute largest power of two divisor of denominator.
         // Always >= 1.
         unchecked {
-            uint twos = (type(uint).max - denominator + 1) & denominator;
+            uint256 twos = (type(uint256).max - denominator + 1) & denominator;
             // Divide denominator by power of two
             assembly {
                 denominator := div(denominator, twos)
@@ -210,7 +210,7 @@ library LibMath {
             // modulo 2**256 such that denominator * inv = 1 mod 2**256.
             // Compute the inverse by starting with a seed that is correct
             // correct for four bits. That is, denominator * inv = 1 mod 2**4
-            uint inv = (3 * denominator) ^ 2;
+            uint256 inv = (3 * denominator) ^ 2;
             // Now use Newton-Raphson iteration to improve the precision.
             // Thanks to Hensel's lifting lemma, this also works in modular
             // arithmetic, doubling the correct bits in each step.
