+++ test/Well.FeeOnTransfer.t.sol
@@ -12,8 +12,8 @@ import {IWellErrors} from "src/interfaces/IWellErrors.sol";
  * (swapFrom, swapTo, addLiquidity) should revert if there actually are fees.
  */
 contract WellFeeOnTransferTest is TestHelper {
-    event AddLiquidity(uint[] tokenAmountsIn, uint lpAmountOut, address recipient);
-    event RemoveLiquidity(uint lpAmountIn, uint[] tokenAmountsOut, address recipient);
+    event AddLiquidity(uint256[] tokenAmountsIn, uint256 lpAmountOut, address recipient);
+    event RemoveLiquidity(uint256 lpAmountIn, uint256[] tokenAmountsOut, address recipient);
 
     function setUp() public {
         setupWellWithFeeOnTransfer(2);
@@ -21,30 +21,30 @@ contract WellFeeOnTransferTest is TestHelper {
     }
 
     function test_swapTo_feeOnTransfer() public prank(user) {
-        uint minAmountOut = 500 * 1e18;
-        uint amountIn = 1000 * 1e18;
+        uint256 minAmountOut = 500 * 1e18;
+        uint256 amountIn = 1000 * 1e18;
 
         vm.expectRevert(IWellErrors.InvalidReserves.selector);
-        well.swapTo(tokens[0], tokens[1], amountIn, minAmountOut, user, type(uint).max);
+        well.swapTo(tokens[0], tokens[1], amountIn, minAmountOut, user, type(uint256).max);
     }
 
     function test_swapFrom_feeOnTransfer() public prank(user) {
-        uint minAmountOut = 500 * 1e18;
-        uint amountIn = 1000 * 1e18;
+        uint256 minAmountOut = 500 * 1e18;
+        uint256 amountIn = 1000 * 1e18;
 
         vm.expectRevert(IWellErrors.InvalidReserves.selector);
-        well.swapFrom(tokens[0], tokens[1], amountIn, minAmountOut, user, type(uint).max);
+        well.swapFrom(tokens[0], tokens[1], amountIn, minAmountOut, user, type(uint256).max);
     }
 
     function test_addLiquidity_feeOnTransfer() public prank(user) {
-        uint[] memory amounts = new uint[](tokens.length);
+        uint256[] memory amounts = new uint256[](tokens.length);
 
-        for (uint i = 0; i < tokens.length; i++) {
+        for (uint256 i = 0; i < tokens.length; i++) {
             amounts[i] = 1000 * 1e18;
         }
-        uint lpAmountOut = 1000 * 1e24;
+        uint256 lpAmountOut = 1000 * 1e24;
 
         vm.expectRevert(IWellErrors.InvalidReserves.selector);
-        well.addLiquidity(amounts, lpAmountOut, user, type(uint).max);
+        well.addLiquidity(amounts, lpAmountOut, user, type(uint256).max);
     }
 }
